Week9

+++ BookShop/controller/LikeController.js
@@ -1,31 +1,67 @@
+const jwt = require('jsonwebtoken')
 const conn = require('../mariadb');
 const {StatusCodes} = require('http-status-codes');
+const dotenv = require('dotenv');
+dotenv.config()
+
+function ensureAuthorization(req, res) {
+
+    try {
+        let receivedJwt = req.headers["authorization"];
+        let decodedJwt = jwt.verify(receivedJwt, process.env.PRIVATE_KEY);
+        return decodedJwt
+    } catch(err) {
+        return err
+    }
+
+}
+
 
 const addLike = (req,res) => {
 
-    const {id} = req.params;
-    const {user_id} = req.body;
+    const book_id = req.params.id;
 
+    let authorization = ensureAuthorization(req, res)
 
-    let sql = "Insert Into likes (user_id, liked_book_id) values (?, ?)"
-    let values = [user_id, id]
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        })
+    } else if (authorization instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else {
+        let sql = "Insert Into likes (user_id, liked_book_id) values (?, ?)"
+        let values = [authorization.id, book_id]
 
-    conn.query(sql, values, (err, results) => {
-        if (err) {
-            return res.status(StatusCodes.BAD_REQUEST).end();
-        }
+        conn.query(sql, values, (err, results) => {
+            if (err) {
+                return res.status(StatusCodes.BAD_REQUEST).end();
+            }
 
-        return res.status(StatusCodes.OK).json(results);
-    })
+            return res.status(StatusCodes.OK).json(results);
+        })
+    }
+
+    
 }
 
 const removeLike = (req, res) => {
-    const {id} = req.params;
-    const {user_id} = req.body;
+    const book_id = req.params.id;
+    let authorization = ensureAuthorization(req, res)
 
-
-    let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?"
-    let values = [user_id, id]
+    if (authorization instanceof jwt.TokenExpiredError) {
+        return res.status(StatusCodes.UNAUTHORIZED).json({
+            "message" : "로그인 세션이 만료되었습니다. 다시 로그인 하세요."
+        })
+    } else if (authorization instanceof jwt.JsonWebTokenError){
+        return res.status(StatusCodes.BAD_REQUEST).json({
+            "message" : "잘못된 토큰입니다."
+        });
+    } else {
+        let sql = "DELETE FROM likes WHERE user_id = ? AND liked_book_id = ?"
+    let values = [authorization.id, book_id]
 
     conn.query(sql, values, (err, results) => {
         if (err) {
@@ -34,7 +70,12 @@ const removeLike = (req, res) => {
 
         return res.status(StatusCodes.OK).json(results);
     })
+    }
+
+
+    
 }
+
 
 module.exports = {
     addLike,
